Video playback reporting ad close before dismissing ad
The bug we experienced had to do with presenting a view controller after closing the interstitial ad unit. The problem was that before the ad was dismissed AA was notifying MoPub that the ad was already successfully closed when it was not. This caused our own presentation to fail. Adding a completion block to the dismiss call provides proper notification when the ad has been successfully closed.